권한이 있는 사람에게만, 수정, 삭제 버튼 노출

+++ src/webapp/article/detail.jsp
@@ -53,8 +53,12 @@
     </div>
 
     <div class="btns flex justify-end gap-x-[5px] mt-3">
-      <a class="btn btn-primary" href="doDelete?id=${param.id}">삭제</a>
-      <a class="btn btn-secondary" href="modify?id=${param.id}">수정</a>
+      <c:if test="${article.extra__actorCanModify}">
+        <a class="btn btn-secondary" href="modify?id=${param.id}">수정</a>
+      </c:if>
+      <c:if test="${article.extra__actorCanDelete}">
+        <a onclick="if( !confirm('정말로 삭제하시겠습니까?') ) return false;" class="btn btn-primary" href="doDelete?id=${param.id}">삭제</a>
+      </c:if>
     </div>
   </div>
 </section>
